Add cron job to migrate videos to OpenVeo
A cron job has been added to migrate Moodle video
files to OpenVeo Publish. Cron is launched every
day at midnight by default. Cron won't do anything
at install as automatic migration is deactivated
by default (can be activated in settings page).
If automatic migration is deactivated only videos
listed in "tool_openveo_migration" table with a
"planned" status will be migrated when the cron
task is executed. If automatic migration is
activated, Moodle video files will be
automatically added to planned videos when cron
task is executed.

"tool_openveo_migration" table keeps trace of the
planned, migrating and migrated videos.

Migration is implemented as a finite states
machine. If migrating a video failed, the states
machine will roll back modifications to the first
state. Then the state of the video will be the
first state of the machine. If something goes
wrong during the rollback then state will be the
current state in the rollback process. Videos on
error won't be automatically retried. Videos in
error with a state different from the first state
must not be retried without human intervention,
it means that rollback process didn't work thus
video might be in an unstable condition.

If something goes wrong in the migration process,
a rollback mechanism will restore the Moodle video
file just like it was and file will be marked as
on "error" in tool_openveo_migration" table.

Migration steps are logged into Moodle logs
storage.