Add planning page
If automatic migration setting was disabled it
wasn't possible to migrate videos. A planning page
has been added to be able to select videos that
will be migrated on next task execution. Planning
page is also a good place to get information of
what is going on, which videos are being migrated,
which videos have been migrated, which videos
are in error.