Add the possibility to not choose a video platform
In configuration page, the first video platform,
in the list of platforms returned by OpenVeo, was
selected by default when displaying the page but
it wasn't saved. As no platform is saved as the
destination platform by default, the
configuration page now won't display a selected
platform anymore if not saved.