Improve names and descriptions of migrated videos
The names of migrated videos on OpenVeo were
containing only the original video file name while
the descriptions didn't contain anything.

Names and descriptions of migrated videos on
OpenVeo can now provide more information about the
video. The description now list the original video
and all its aliases with a link to the
corresponding Moodle page.

The name is now customizable through the settings
page. The name format can be specified by context
type (module, block, course, category and user).
The name format accepts tokens that will be
replaced by context information at the time of
migration. The resulting text is used as the
migrated video name but also in the description
for the list of aliases.