restore_original: make transition more resilient
restore_original transition was failing if next
transition started to create new references.
Moodle couldn't re-create the original file entry
in files table as pathnamehash property is unique.

Instead of supposing that the file is in the trash
directory (it might not as Moodle won't remove the
file if a reference still exists in the files
table) we use
file_system->get_local_path_from_storedfile to
look for the path of the file if in Moodle data
and automatically recover the file if in trash
directory.